Start adding some utility functions for gia demo
